pyproject.toml
[tool.poetry]
name = "python_deep_learning_framework_stars"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^1.55"
Jinja2 = "^3.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^21.8b0"
flake8 = "^3.9.2"
ipython = "^7.27.0"
jedi = "^0.18.0"
mypy = "^0.910"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
